Test Boss Prep
Actual enemy and code tomorrow or something. Just taking the first steps to prepare for today. The base texture is done, although the boss is lacking all kinds of attack textures and a shield texture - I will add these later. Regarding the music I intend to use, it's a slightly modified version of a royalty free soundtrack by Kevin MacLeod. All credit to him.

I'll be honest this seems like a colossal task and I'll probably add attacks as I go and add more functionality to the best of my ability.