Reworked the Greenstrike Dynamo (Now Lifeforce Engine)
Oh boy, haven't actually done a lot of work on this mod in a long time. Feels good to finally have made something that works.  It may not have pretty visuals or sound design yet, but the functionality is there and it's exactly what I was hoping for. This may also be subjected to severe balance changes since it is literally super overpowered at the moment but only time will tell.